# Registro-Modernozip/Registro-Moderno/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Provider {
  id: number;
  name: string;
  rg: string;
  company: string;
  defaultService: string;
  /** @nullable */
  photoData: string | null;
  createdAt: string;
  /** @nullable */
  lastVisitAt: string | null;
  visitCount: number;
}

export interface ProviderInput {
  /** @minLength 2 */
  name: string;
  /** @minLength 3 */
  rg: string;
  /** @minLength 2 */
  company: string;
  /** @minLength 2 */
  defaultService: string;
  /** @nullable */
  photoData?: string | null;
}

export interface ProviderUpdate {
  /** @minLength 2 */
  name?: string;
  /** @minLength 3 */
  rg?: string;
  /** @minLength 2 */
  company?: string;
  /** @minLength 2 */
  defaultService?: string;
  /** @nullable */
  photoData?: string | null;
}

export interface Visit {
  id: number;
  providerId: number;
  providerName: string;
  company: string;
  service: string;
  /** @nullable */
  photoData: string | null;
  enteredAt: string;
  /** @nullable */
  exitAt: string | null;
}

export interface VisitInput {
  /** @minimum 1 */
  providerId: number;
  /** @minLength 2 */
  service: string;
  /** @nullable */
  exitAt?: string | null;
}

export interface VisitExitInput {
  /** @nullable */
  exitAt: string | null;
}

export interface DashboardSummary {
  totalProviders: number;
  visitsToday: number;
  currentlyExpected: number;
  recentVisits: Visit[];
}

export interface RequestUploadInput {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata: RequestUploadInput;
}

export type ListProvidersParams = {
search?: string;
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
};

export type ListVisitsParams = {
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
};