# Registro-Modernozip/Registro-Moderno/lib/api-zod/src/generated/types/visit.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Visit {
  id: number;
  providerId: number;
  providerName: string;
  company: string;
  service: string;
  /** @nullable */
  photoData: string | null;
  enteredAt: Date;
  /** @nullable */
  exitAt: Date | null;
}